chore: move base test classes into skill for easier reuse
