+++ docs/quick_start.md
@@ -86,7 +86,7 @@ In this case, we consider the [Hello World agent service](./hello_world_agent_se
 
 2. Use the CLI to download and build the images to deploy the [Hello World agent service](./hello_world_agent_service.md):
     ```bash
-    autonomy deploy build deployment valory/hello_world:0.1.0:bafybeicqjg7jd2ea3uzasw6xikfqojypsdxi5r5776ydzxyyyi5epa54mm keys.json --remote
+    autonomy deploy build deployment valory/hello_world:0.1.0:bafybeigwc5snwpge5qvzhwo7mzzezcopjld6izqwp7lwaitvzvsho73way keys.json --remote
     ```
     The command above generates the required images to run the agent service using the keys provided in the `keys.json` file. In this case, we are accessing the service definition located in the Service Registry.
