+++ packages/hashes.csv
@@ -4,16 +4,16 @@ open_aea/protocols/scaffold,bafybeihhyjuqvp5zqtikkjrwtikg4fr4tsub43n777huxtpeu77
 open_aea/protocols/signing,bafybeieirkuq6azgis33bmukr5u2gx4chp5kswiw3wd5hbjsaukgwzmp5a
 open_aea/skills/scaffold,bafybeih3h7yia76rrbittnoeotpraiuho5rdqsnksuvncinaa3akgmnbaa
 valory/agents/abstract_abci,bafybeidxzirfyeshqq4nvvm2cqd4xou2e7aent7zg33tgcji6yvovodlqu
-valory/agents/apy_estimation,bafybeihvvm36zukty2u4uy2efgiqtqsb2ul2bl3bjxms5jhiyazb3vrhmi
-valory/agents/apy_estimation_chained,bafybeiaxfufmirxf6texozstzefyqvoov2iqzm42hng7n2e3k7z5p5b7ai
+valory/agents/apy_estimation,bafybeih2dndxnbhwcrtwuqfq4x43h7bxrlb3s7zkyvboa4lwircselih4u
+valory/agents/apy_estimation_chained,bafybeiac72etm4sd446wzpcyywi44mzumf3yocqh74edknbjwl2g6d3pry
 valory/agents/counter,bafybeiai6quagduazo2spo3eylho3qrrqgxf3c7m7l54llksbhgcst2gue
 valory/agents/counter_client,bafybeigonx77seeshstpz2nhyx4gaxddtjjty4toxuh3wvqwh4i2etkire
-valory/agents/hello_world,bafybeig4clacvvxywtowgdflmoib35lrq2jpgwjtbue632n2reclov2lhm
-valory/agents/oracle,bafybeibfse7av7x2q4gwh5owestqbxog6fjn72etraf5tjgemrdrf42lfu
-valory/agents/register_reset,bafybeifxkfg4qocnfjlcweq7j3adxgpxewxasmkpodpanhhsqxemb43xoa
-valory/agents/registration_start_up,bafybeicj4yhuqh7hgr2dgvrm3b5kudejtnlprrspbzipffjwfj3fvvga4q
-valory/agents/simple_abci,bafybeietwlgphdi4kdf2taido63c26w5q26hjjtpmgkyzpjp74otkzg3qm
-valory/agents/test_abci,bafybeiabyk3uaz6ud7p2lsajk3ns6cmfyez7w4igqxlqwedcwz77oeu2n4
+valory/agents/hello_world,bafybeia7j55y6bxtasr5h7digjflzaoiao32secgpwvbyasffneilw7jcm
+valory/agents/oracle,bafybeiablf55fw7w7xrdx4ulr2a4i7bm3yiedmebgyjqcphakbh2w7h2z4
+valory/agents/register_reset,bafybeibzuhprhc5jt4fpzimfwhhyymjlqqtfpdzzpt6j53vohh4spo6o3u
+valory/agents/registration_start_up,bafybeidefmhorgyta3uswkf2tfkab2vjcgsjr5wx5zl4swye2ugi5hswru
+valory/agents/simple_abci,bafybeiho76ux7ovd2xu5ouhvscfq67r5kkmfiabnnaaomskbos65mwly7a
+valory/agents/test_abci,bafybeidiifw45dc2zbnxkrfno34vy7v6bjdby26sdulc5rtzibrayfkoh4
 valory/connections/abci,bafybeibfdzfevrtiywolqeepk6yfzyvl4bbu6mghhnebkptcbhsh5fxkbu
 valory/connections/http_client,bafybeibh3iu7kojc4d5eiigqh6kx2vdmu2q7vhcvvq2rz6suegsscyd66y
 valory/connections/ledger,bafybeifxxshi4u3xyyllip5e3yafawywcwpvqxgnybchi5kj74zm4b4n2e
@@ -31,31 +31,31 @@ valory/protocols/contract_api,bafybeie5gbzc4v4n6p6ea66kmc5ejwvbjzhd7vhcjg7mikbyx
 valory/protocols/http,bafybeifhmibx3l7f3xdwxhweczizadxs6p3oludvapvwdeu4o5f3oahqme
 valory/protocols/ledger_api,bafybeifaozy766hngsqc5a5kpebseusq4cc3woh4ayycdoa7r3mjkvhzr4
 valory/protocols/tendermint,bafybeigrenadbkdzpkxbjmpkubxzmihk45v6jbwosngciqdccth4lvp7cm
-valory/services/apy_estimation_hardhat,bafybeiab62tdobex2wl77xjakkbceynoau5bgjkdmumwjl3jvrjeojp5h4
+valory/services/apy_estimation_hardhat,bafybeidw3kok46r7o7byozzywg3s7y437rn4m4irwsgfxhoxh2q52hbzqe
 valory/services/counter,bafybeieyxybssaoirqo3apmarxt7bly3ci5uzj43npfrehbxzljwcmu63u
-valory/services/hello_world,bafybeicqjg7jd2ea3uzasw6xikfqojypsdxi5r5776ydzxyyyi5epa54mm
-valory/services/oracle_goerli,bafybeidfxckcqrleyx5biylf7qdavwpgpicjna5snchdep4jqcnxwsqdvy
-valory/services/oracle_hardhat,bafybeiasqsov5o7woltsktlsrs7nk2qnocq4aun7otujbjhlpxnf6a3lzy
-valory/services/oracle_polygon,bafybeidkpfgcqg5wotugujhjpadtftmw7e5ixblomac7uyidjvneerowau
-valory/services/oracle_ropsten,bafybeic6kdvy4b5bvhylfmy3cd6pirimlh7qwqex2jb7fqbuyihiksjwza
-valory/services/oracle_ropsten_cluster,bafybeigpjjd2iw2bhkqib4y6ot2zgmt6w5s2mvons53gauxo6uvyakzqfq
-valory/services/register_reset,bafybeiavsy4vkuhhgha67zgoo6w4gbrygheo7lyf3k2feai5ml4btt4nay
+valory/services/hello_world,bafybeigwc5snwpge5qvzhwo7mzzezcopjld6izqwp7lwaitvzvsho73way
+valory/services/oracle_goerli,bafybeidr75742rbnboqvzd6jq6driklinwdt73swssq2gry7xkkbayu7ue
+valory/services/oracle_hardhat,bafybeigyzugvoh36bh5r2l4fxuwvf6zqxmsckg5fxjdrxpbmpszzy4oxdi
+valory/services/oracle_polygon,bafybeigiul54r2i77cybyyfhxro6jrbt5w2s7kpaaaeko266qe7bekbzzi
+valory/services/oracle_ropsten,bafybeihbcjurdpgrn6sh5hyio7xkbu3sbobnoigoui2qsl4qimh5yhopsa
+valory/services/oracle_ropsten_cluster,bafybeiegbgaep6fxtvu6pnhftqmlyvlwnomxgy3putj6mqrmcsr6md2kd4
+valory/services/register_reset,bafybeiapcexdp4ghrznlz7xbbwkn57kgfyhze2zziqrvoxapvycaats6xm
 valory/skills/abstract_abci,bafybeie2cs5y7xu625birt7i6vlrhakhody2k4txf7soef3lqwhexpfgqy
-valory/skills/abstract_round_abci,bafybeihz4pkaqwufq2szqr3ji3kjsbjodsg5yultbit3xdcfjdwi4ebgoa
-valory/skills/apy_estimation_abci,bafybeifinpdbjocz5wy6j4mzdclauqoxko54l7ve2ewsm5rq4jrgrm5wei
-valory/skills/apy_estimation_chained_abci,bafybeiaowwuv55lofj6op2ygcl6u2cz2q3wcgwf6hvke6xbljzliz2mc7e
+valory/skills/abstract_round_abci,bafybeidjptu3fnsna3ei2kudxbpydejgkrvw2csrwy55wu5in57fqwgnlm
+valory/skills/apy_estimation_abci,bafybeiajxhv23qvekodjqxy2e5fdiwldhgb7iukmfb6nkafvc5jbn2gxne
+valory/skills/apy_estimation_chained_abci,bafybeiamadd2x7osjvbzrr3ggmzzxbwte5pijgxno6lbojm2kkphdv7uqy
 valory/skills/counter,bafybeigsjhaj7bii2t3psorw7wgyr43xbavh2wxeecphdm3bmuv2w63evm
 valory/skills/counter_client,bafybeiggq3dazj2on553qv35sjotqli7kn7o5icedlowo7zwx7amxonwga
-valory/skills/hello_world_abci,bafybeibxzheclohkujv3g2ily5fho6sg4rywbt4lvkofqtz27n3tdl4nv4
-valory/skills/liquidity_provision_abci,bafybeiedexyek5j5lljbwklxcxd4mi7fobeup7ozsq67qfzgvqxp4zip5q
-valory/skills/liquidity_rebalancing_abci,bafybeiciiaptkdxgwjjxetvnw6ugcae5kp3k3orzqz52dfccjejijylhxy
-valory/skills/oracle_abci,bafybeih2xje4nrndmk25y33nkijbpyohsv4tzcrycjvgk54g7oy4m6p7yq
-valory/skills/oracle_deployment_abci,bafybeie6h6s7e7zcggzfbcnbf64sqaihjf5d64hbamfn5t7lckv3pv7awy
-valory/skills/price_estimation_abci,bafybeicbzshlxz2n7mscelzum3inz4fw5xhrulm56siv3cvr427v7c4bhe
-valory/skills/register_reset_abci,bafybeihbqapa6b2xz73yidtqar3oobrib6cz4t5u24sopdulbk6cop6bxi
-valory/skills/registration_abci,bafybeiamr2gckrgxujl6xskhtjg46fmdwsklq5fnqvndnwsiyld6sy5z4y
-valory/skills/reset_pause_abci,bafybeibhqdl3263vcokyjxeciiktbk2ej6gdbeexviy5exvnc37a5hdmhq
-valory/skills/safe_deployment_abci,bafybeiaqnl5az4fdkv7yyjg6s4rouhpldkgpplmohqdt34wwawfkj5jeva
-valory/skills/simple_abci,bafybeia6isegqmlwfx2rjpmrqg4jm63zkacdgtyo4mzwgeuw3yzq4jsjau
-valory/skills/test_abci,bafybeig6qiq24xstl42inaklbdheut2m4dm62lasvajxrxqwesq5w2njgy
-valory/skills/transaction_settlement_abci,bafybeic6khhxyes6kdigczfuxqhr6tcujuj2ukwhlgb3pt3kr4cpmnygaq
+valory/skills/hello_world_abci,bafybeieef4lgfau3gisrpko2gq2y43scieworylxynfnbalxmsqyzpzzhu
+valory/skills/liquidity_provision_abci,bafybeid7hwrdmqxpnq2ytdqcklgf7bgwsmm5377diwxcdg4fhqbp4myevu
+valory/skills/liquidity_rebalancing_abci,bafybeihboeydxefv2y2eubidxjzatscqvkslwwun75zdgkcvz5cy37nvn4
+valory/skills/oracle_abci,bafybeifzniprzvcqu4uwmgrmgxtuuv72qkc4eil5mcjrqewa7pril4ar64
+valory/skills/oracle_deployment_abci,bafybeie3krcpijnymhsr27o6yd7wm2224xx5b53w5qagqhuboie5pnub4i
+valory/skills/price_estimation_abci,bafybeicms57fg4igxwkbdqkqoanwjtrtd4hqenjhkp7n2ayreysluswbgq
+valory/skills/register_reset_abci,bafybeicjkug27zogozlupgncaux7jj55rzjg2rlz34rk76kqpkxjthdbyi
+valory/skills/registration_abci,bafybeiew7zgxmgoegyuxktoa6bzxgu4ae7kttry76uxcqqpptk3gfmwf4a
+valory/skills/reset_pause_abci,bafybeidagost3f2c7y3l6jh4ocgtoel4bvzt5pmfzboqf3cuq6cjndccuy
+valory/skills/safe_deployment_abci,bafybeic32f4w6fk5bwrrwuwye6pmmqn4eagxwefup7ai2jxfbv5ox77xa4
+valory/skills/simple_abci,bafybeif5vgvu3p2zybdvg4fwvh74q3fsqs5eqhiwu6djrakszrfbpbjzda
+valory/skills/test_abci,bafybeicudtm5lb6msbztm4cyzkmuzoehh3h5bilugl6wkdqdoqg72gd4qq
+valory/skills/transaction_settlement_abci,bafybeigr4zhffjvxbkod534kdtpv3wforr4j6wo7gebbp7w35gphzmatve
