+++ packages/valory/agents/apy_estimation_chained/aea-config.yaml
@@ -24,10 +24,10 @@ protocols:
 - valory/tendermint:0.1.0:bafybeigrenadbkdzpkxbjmpkubxzmihk45v6jbwosngciqdccth4lvp7cm
 skills:
 - valory/abstract_abci:0.1.0:bafybeie2cs5y7xu625birt7i6vlrhakhody2k4txf7soef3lqwhexpfgqy
-- valory/abstract_round_abci:0.1.0:bafybeihz4pkaqwufq2szqr3ji3kjsbjodsg5yultbit3xdcfjdwi4ebgoa
-- valory/apy_estimation_abci:0.1.0:bafybeifinpdbjocz5wy6j4mzdclauqoxko54l7ve2ewsm5rq4jrgrm5wei
-- valory/apy_estimation_chained_abci:0.1.0:bafybeiaowwuv55lofj6op2ygcl6u2cz2q3wcgwf6hvke6xbljzliz2mc7e
-- valory/registration_abci:0.1.0:bafybeiamr2gckrgxujl6xskhtjg46fmdwsklq5fnqvndnwsiyld6sy5z4y
+- valory/abstract_round_abci:0.1.0:bafybeidjptu3fnsna3ei2kudxbpydejgkrvw2csrwy55wu5in57fqwgnlm
+- valory/apy_estimation_abci:0.1.0:bafybeiajxhv23qvekodjqxy2e5fdiwldhgb7iukmfb6nkafvc5jbn2gxne
+- valory/apy_estimation_chained_abci:0.1.0:bafybeiamadd2x7osjvbzrr3ggmzzxbwte5pijgxno6lbojm2kkphdv7uqy
+- valory/registration_abci:0.1.0:bafybeiew7zgxmgoegyuxktoa6bzxgu4ae7kttry76uxcqqpptk3gfmwf4a
 default_ledger: ethereum
 required_ledgers:
 - ethereum
