+++ packages/valory/agents/hello_world/aea-config.yaml
@@ -18,8 +18,8 @@ protocols:
 - valory/http:1.0.0:bafybeifhmibx3l7f3xdwxhweczizadxs6p3oludvapvwdeu4o5f3oahqme
 skills:
 - valory/abstract_abci:0.1.0:bafybeie2cs5y7xu625birt7i6vlrhakhody2k4txf7soef3lqwhexpfgqy
-- valory/abstract_round_abci:0.1.0:bafybeihz4pkaqwufq2szqr3ji3kjsbjodsg5yultbit3xdcfjdwi4ebgoa
-- valory/hello_world_abci:0.1.0:bafybeibxzheclohkujv3g2ily5fho6sg4rywbt4lvkofqtz27n3tdl4nv4
+- valory/abstract_round_abci:0.1.0:bafybeidjptu3fnsna3ei2kudxbpydejgkrvw2csrwy55wu5in57fqwgnlm
+- valory/hello_world_abci:0.1.0:bafybeieef4lgfau3gisrpko2gq2y43scieworylxynfnbalxmsqyzpzzhu
 default_ledger: ethereum
 required_ledgers:
 - ethereum
