+++ packages/valory/agents/register_reset/aea-config.yaml
@@ -19,10 +19,10 @@ protocols:
 - valory/http:1.0.0:bafybeifhmibx3l7f3xdwxhweczizadxs6p3oludvapvwdeu4o5f3oahqme
 skills:
 - valory/abstract_abci:0.1.0:bafybeie2cs5y7xu625birt7i6vlrhakhody2k4txf7soef3lqwhexpfgqy
-- valory/abstract_round_abci:0.1.0:bafybeihz4pkaqwufq2szqr3ji3kjsbjodsg5yultbit3xdcfjdwi4ebgoa
-- valory/register_reset_abci:0.1.0:bafybeihbqapa6b2xz73yidtqar3oobrib6cz4t5u24sopdulbk6cop6bxi
-- valory/registration_abci:0.1.0:bafybeiamr2gckrgxujl6xskhtjg46fmdwsklq5fnqvndnwsiyld6sy5z4y
-- valory/reset_pause_abci:0.1.0:bafybeibhqdl3263vcokyjxeciiktbk2ej6gdbeexviy5exvnc37a5hdmhq
+- valory/abstract_round_abci:0.1.0:bafybeidjptu3fnsna3ei2kudxbpydejgkrvw2csrwy55wu5in57fqwgnlm
+- valory/register_reset_abci:0.1.0:bafybeicjkug27zogozlupgncaux7jj55rzjg2rlz34rk76kqpkxjthdbyi
+- valory/registration_abci:0.1.0:bafybeiew7zgxmgoegyuxktoa6bzxgu4ae7kttry76uxcqqpptk3gfmwf4a
+- valory/reset_pause_abci:0.1.0:bafybeidagost3f2c7y3l6jh4ocgtoel4bvzt5pmfzboqf3cuq6cjndccuy
 default_ledger: ethereum
 required_ledgers:
 - ethereum
