+++ packages/valory/agents/simple_abci/aea-config.yaml
@@ -21,8 +21,8 @@ protocols:
 - valory/ledger_api:1.0.0:bafybeifaozy766hngsqc5a5kpebseusq4cc3woh4ayycdoa7r3mjkvhzr4
 skills:
 - valory/abstract_abci:0.1.0:bafybeie2cs5y7xu625birt7i6vlrhakhody2k4txf7soef3lqwhexpfgqy
-- valory/abstract_round_abci:0.1.0:bafybeihz4pkaqwufq2szqr3ji3kjsbjodsg5yultbit3xdcfjdwi4ebgoa
-- valory/simple_abci:0.1.0:bafybeia6isegqmlwfx2rjpmrqg4jm63zkacdgtyo4mzwgeuw3yzq4jsjau
+- valory/abstract_round_abci:0.1.0:bafybeidjptu3fnsna3ei2kudxbpydejgkrvw2csrwy55wu5in57fqwgnlm
+- valory/simple_abci:0.1.0:bafybeif5vgvu3p2zybdvg4fwvh74q3fsqs5eqhiwu6djrakszrfbpbjzda
 default_ledger: ethereum
 required_ledgers:
 - ethereum
