+++ packages/valory/agents/test_abci/aea-config.yaml
@@ -21,8 +21,8 @@ protocols:
 - valory/ledger_api:1.0.0:bafybeifaozy766hngsqc5a5kpebseusq4cc3woh4ayycdoa7r3mjkvhzr4
 skills:
 - valory/abstract_abci:0.1.0:bafybeie2cs5y7xu625birt7i6vlrhakhody2k4txf7soef3lqwhexpfgqy
-- valory/abstract_round_abci:0.1.0:bafybeihz4pkaqwufq2szqr3ji3kjsbjodsg5yultbit3xdcfjdwi4ebgoa
-- valory/test_abci:0.1.0:bafybeig6qiq24xstl42inaklbdheut2m4dm62lasvajxrxqwesq5w2njgy
+- valory/abstract_round_abci:0.1.0:bafybeidjptu3fnsna3ei2kudxbpydejgkrvw2csrwy55wu5in57fqwgnlm
+- valory/test_abci:0.1.0:bafybeicudtm5lb6msbztm4cyzkmuzoehh3h5bilugl6wkdqdoqg72gd4qq
 default_ledger: ethereum
 required_ledgers:
 - ethereum
