+++ packages/valory/services/apy_estimation_hardhat/service.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeibqyfur7wusrbfga66hcbauxvhs2g3j5yhknsxlvyh6mil57aajyu
 fingerprint_ignore_patterns: []
-agent: valory/apy_estimation:0.1.0:bafybeihvvm36zukty2u4uy2efgiqtqsb2ul2bl3bjxms5jhiyazb3vrhmi
+agent: valory/apy_estimation:0.1.0:bafybeih2dndxnbhwcrtwuqfq4x43h7bxrlb3s7zkyvboa4lwircselih4u
 network: hardhat
 number_of_agents: 4
 ---
