+++ packages/valory/services/oracle_polygon/service.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeifqnvekc6s44wsczofdhofiozxekstjau2bk5yqfqvjuv4wzoopku
 fingerprint_ignore_patterns: []
-agent: valory/oracle:0.1.0:bafybeibfse7av7x2q4gwh5owestqbxog6fjn72etraf5tjgemrdrf42lfu
+agent: valory/oracle:0.1.0:bafybeiablf55fw7w7xrdx4ulr2a4i7bm3yiedmebgyjqcphakbh2w7h2z4
 network: polygon
 number_of_agents: 4
 ---
