+++ packages/valory/services/register_reset/service.yaml
@@ -1,7 +1,7 @@
 name: register_reset
 author: valory
 version: 0.1.0
-agent: valory/register_reset:0.1.0:bafybeifxkfg4qocnfjlcweq7j3adxgpxewxasmkpodpanhhsqxemb43xoa
+agent: valory/register_reset:0.1.0:bafybeibzuhprhc5jt4fpzimfwhhyymjlqqtfpdzzpt6j53vohh4spo6o3u
 network: hardhat
 number_of_agents: 4
 description: Test and debug tendermint reset mechanism.
