+++ packages/valory/skills/abstract_round_abci/test_tools/__init__.py
@@ -17,17 +17,4 @@
 #
 # ------------------------------------------------------------------------------
 
-"""Test the models.py module of the skill."""
-
-from packages.valory.skills.abstract_round_abci.test_tools.base import DummyContext
-from packages.valory.skills.registration_abci.models import SharedState
-
-
-class TestSharedState:
-    """Test SharedState(Model) class."""
-
-    def test_initialization(
-        self,
-    ) -> None:
-        """Test initialization."""
-        SharedState(name="", skill_context=DummyContext())
+"""Tests package for abstract_round_abci derived skills."""
