+++ packages/valory/skills/register_reset_abci/skill.yaml
@@ -24,9 +24,9 @@ protocols:
 - valory/abci:0.1.0:bafybeicbj36urbjkn6y7hrhm57dspyuuvpcnxhr3r6y2cwxammpom633g4
 - valory/http:1.0.0:bafybeifhmibx3l7f3xdwxhweczizadxs6p3oludvapvwdeu4o5f3oahqme
 skills:
-- valory/abstract_round_abci:0.1.0:bafybeihz4pkaqwufq2szqr3ji3kjsbjodsg5yultbit3xdcfjdwi4ebgoa
-- valory/registration_abci:0.1.0:bafybeiamr2gckrgxujl6xskhtjg46fmdwsklq5fnqvndnwsiyld6sy5z4y
-- valory/reset_pause_abci:0.1.0:bafybeibhqdl3263vcokyjxeciiktbk2ej6gdbeexviy5exvnc37a5hdmhq
+- valory/abstract_round_abci:0.1.0:bafybeidjptu3fnsna3ei2kudxbpydejgkrvw2csrwy55wu5in57fqwgnlm
+- valory/registration_abci:0.1.0:bafybeiew7zgxmgoegyuxktoa6bzxgu4ae7kttry76uxcqqpptk3gfmwf4a
+- valory/reset_pause_abci:0.1.0:bafybeidagost3f2c7y3l6jh4ocgtoel4bvzt5pmfzboqf3cuq6cjndccuy
 behaviours:
   main:
     args: {}
