+++ tests/test_skills/test_oracle_deployment_abci/test_behaviours.py
@@ -41,6 +41,9 @@ from packages.valory.skills.abstract_round_abci.behaviour_utils import (
     BaseBehaviour,
     make_degenerate_behaviour,
 )
+from packages.valory.skills.abstract_round_abci.test_tools.base import (
+    FSMBehaviourBaseCase,
+)
 from packages.valory.skills.oracle_deployment_abci.behaviours import (
     DeployOracleBehaviour,
     RandomnessOracleBehaviour,
@@ -58,7 +61,6 @@ from packages.valory.skills.oracle_deployment_abci.rounds import (
 from packages.valory.skills.oracle_deployment_abci.rounds import FinishedOracleRound
 
 from tests.conftest import ROOT_DIR
-from tests.test_skills.base import FSMBehaviourBaseCase
 from tests.test_skills.test_abstract_round_abci.test_common import (
     BaseRandomnessBehaviourTest,
     BaseSelectKeeperBehaviourTest,
