+++ tests/test_skills/test_safe_deployment_abci/test_models.py
@@ -19,9 +19,8 @@
 
 """Test the models.py module of the skill."""
 
+from packages.valory.skills.abstract_round_abci.test_tools.base import DummyContext
 from packages.valory.skills.safe_deployment_abci.models import SharedState
-
-from tests.test_skills.base import DummyContext
 
 
 class TestSharedState:
